mocking and customerrors

+++ src/dao/repositories/products.repository.js
@@ -2,9 +2,33 @@ const productsModel = require('../models/products');
 const ProductDTO = require('../dto/products.dto');
 const fs = require('fs');
 const { transformDataProducts } = require('../../utils/transformdata');
+const { faker } = require("@faker-js/faker")
 
 class ProductsRepository {
     constructor() { }
+
+    async getMockingProducts() {
+        try {
+            const mockingProducts = [];
+            for (let i = 0; i < 100; i++) {
+                const product = await productsModel.create({
+                    title: faker.commerce.productName(),
+                    description: faker.commerce.productDescription(),
+                    code: faker.string.numeric(5),
+                    price: faker.commerce.price(),
+                    stock: faker.string.numeric(),
+                    category: faker.commerce.department(),
+                    thumbnail: faker.image.url()
+                })
+                mockingProducts.push(product)
+            }
+            console.log(mockingProducts.length)
+            return mockingProducts
+        } catch (error) {
+            console.log(error)
+        }
+    }
+
 
     async paginateProducts(category, status, limit, sort, page, limitQueryParams, order) {
         let query = {};
@@ -23,6 +47,9 @@ class ProductsRepository {
 
     async getProducts() {
         const products = await productsModel.find();
+        if (!products) {
+            return { status: 404, error: 'Error try found products', answer: products };
+        }
         const dataProducts = transformDataProducts(products);
         return dataProducts;
     }
@@ -30,7 +57,7 @@ class ProductsRepository {
     async getProductById(pid) {
         const product = await productsModel.findById(pid);
         if (!product) {
-            return { error: `The product with ID ${pid} doesn't exist` };
+            return { error: `The product with ID ${pid} doesn't exist`, answer: product  };
         }
         return product;
     }
@@ -41,6 +68,9 @@ class ProductsRepository {
         }
         const productdto = new ProductDTO(body);
         const product = await productsModel.create(productdto);
+        if (!product) {
+            return { error: `Error trying create product`, answer: product };
+        }
         return product;
     }
 
@@ -52,6 +82,9 @@ class ProductsRepository {
 
         if (file) {
             const product = await productsModel.findById(pid);
+            if (!product) {
+                return { error: `Error trying find product`, answer: product };
+            }
             if (product.thumbnail !== 'file') {
                 const nameFile = product.thumbnail.split("/").pop();
                 fs.unlinkSync(`${__dirname}/../../public/storage/products/${nameFile}`);
@@ -60,11 +93,17 @@ class ProductsRepository {
 
         const productdto = new ProductDTO(dataReplace);
         const productReplaced = await productsModel.findByIdAndUpdate(pid, productdto, { new: true });
+        if (!productReplaced) {
+            return { error: `Error trying update product`, answer: productReplaced };
+        }
         return productReplaced;
     }
 
     async deleteProductById(id) {
         const product = await productsModel.findByIdAndDelete(id);
+        if (!product) {
+            return { error: `Error trying delete product`, answer: product };
+        }
         if (product.thumbnail !== 'file') {
             const file = product.thumbnail.split("/").pop();
             if (file) {
@@ -72,9 +111,6 @@ class ProductsRepository {
             }
         } else {
             console.log("Product has no thumbnail.");
-        }
-        if (!product) {
-            throw { error: `The product with ID ${pid} doesn't exist` };
         }
         return product;
     }
